requirement.txt
mysql-connector-python
py-cord
pillow
python-dotenv
